Rename BalancerPoolCard & PoolSummary component
