+++ src/theme/index.tsx
@@ -48,7 +48,7 @@ export function colors(darkMode: boolean): Colors {
     text5: darkMode ? '#2C2F36' : '#EDEEF2',
 
     // backgrounds / greys
-    bg1: darkMode ? '#18212e' : '#FFFFFF',
+    bg1: darkMode ? '#18212e' : '#FBFAFF',
     bg2: darkMode ? '#202d3f' : '#471BB2',
     bg3: darkMode ? '#2a3a50' : '#F1F1F1',
     bg4: darkMode ? '#3a506f' : '#CED0D9',
